requirements.txt
ultralytics>=8.1.0
scikit-learn
pandas
opencv-python>=4.8.0
numpy>=1.23.0
Pillow>=9.5.0
gradio>=4.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
google-genai
